Add files via upload
We created a skill on the Amazon Alexa developer website. This will allow us to test our project by taking the humidity measurements from the weather report for the day. Since a raspberry pi is not accessible to us at the moment we will act as if Alexa was taking the humidity from the raspberry pi sensor.